# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "description": "Payload of a single API error.\n\nСодержимое поля ``error`` в ответе. Используется FE для\nлокализации и анализа: код (``code``) стабилен и пригоден для\n``switch``-логики, ``message`` — текст для пользователя/лога,\n``fields`` — пер-полевые ошибки валидации.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Fields",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "ApiErrorResponse": {
        "description": "Top-level error envelope returned by the API.\n\nВнешняя «обёртка», чтобы у любого ответа об ошибке был один и тот\nже ключ верхнего уровня. Эта схема также прокидывается в OpenAPI\nкак ``responses`` для роутов, чтобы FE-генератор типов знал форму.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "CurrentUser": {
        "description": "Represents the current authenticated user.\n\nDTO, который возвращает ``GET /auth/me`` и который кладётся в\nзависимостях защищённых роутов. ``alias_generator=to_camel`` нужен,\nчтобы JSON получил ``displayName`` (camelCase), а Python-код мог\nобращаться к ``display_name`` (snake_case).",
        "properties": {
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "id"
        ],
        "title": "CurrentUser",
        "type": "object"
      },
...
      "LogoutRequest": {
        "description": "Request body for idempotent refresh-token based logout.",
        "properties": {
          "refreshToken": {
            "title": "Refreshtoken",
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "title": "LogoutRequest",
        "type": "object"
      },
...
      "OtpRequest": {
        "description": "Request body for starting email OTP login.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "OtpRequest",
        "type": "object"
      },
      "OtpRequestDevResponse": {
        "description": "Dev/local/test OTP response with the raw code for local testing.",
        "properties": {
          "expiresAt": {
            "title": "Expiresat",
            "type": "integer"
          },
          "otp": {
            "title": "Otp",
            "type": "string"
          }
        },
        "required": [
          "expiresAt",
          "otp"
        ],
        "title": "OtpRequestDevResponse",
        "type": "object"
      },
      "OtpVerifyRequest": {
        "description": "Request body for verifying email OTP login.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "otp": {
            "maxLength": 6,
            "minLength": 6,
            "title": "Otp",
            "type": "string"
          }
        },
        "required": [
          "email",
          "otp"
        ],
        "title": "OtpVerifyRequest",
        "type": "object"
      },
      "OtpVerifyResponse": {
        "description": "Successful OTP verification response.",
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "refreshToken": {
            "title": "Refreshtoken",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/CurrentUser"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ],
        "title": "OtpVerifyResponse",
        "type": "object"
      },
      "RefreshRequest": {
        "description": "Request body for rotating a refresh token.",
        "properties": {
          "refreshToken": {
            "title": "Refreshtoken",
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "RefreshResponse": {
        "description": "Successful refresh-token rotation response.",
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "refreshToken": {
            "title": "Refreshtoken",
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ],
        "title": "RefreshResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "MSD Course",
      "url": "https://github.com/larchanka-training/dmc-1-t2-notebook-api"
    },
    "description": "## JS Notebook API\n\nBackend API for the JS Notebook project. Built on a multi-module architecture: each domain module owns its `controllers`, `services` and `schemas`.\n\n**Documentation:**\n- Swagger UI: [`/api/v1/docs`](/api/v1/docs)\n- ReDoc: [`/api/v1/redoc`](/api/v1/redoc)\n- OpenAPI schema: [`/api/v1/openapi.json`](/api/v1/openapi.json)\n",
    "license": {
      "name": "MIT"
    },
    "title": "JS Notebook API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revoke the session identified by the refresh token.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Session revoked or already inactive"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Logout current refresh-token session",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Returns the user owning the Bearer access token.",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get current user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/otp/request": {
      "post": {
        "description": "Create and send an OTP for the provided email address.",
        "operationId": "request_otp_api_v1_auth_otp_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestDevResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "OTP sent without response body"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid email"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Request email OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "description": "Verify an OTP and return access/refresh tokens.",
        "operationId": "verify_otp_api_v1_auth_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpVerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid email"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid or expired OTP"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Verify email OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Rotate a refresh token and return a new token pair.",
        "operationId": "refresh_tokens_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid refresh token"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Refresh auth tokens",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
...
}